Added a generic Ajax submitForm function. Removed saving of form before submit.
Up to now the form was saved in storage despite the user being online
or offline. It is more efficient to try submit the form and only then
if unsuccessful - save it. Also, the Firefox somehow could not cope with
resizing the photos, saving all data in storage and then suddenly submitting
it all. Maybe the submit was too fast as some actions are asynchronous?